AI HR Assistant/requirements.txt
pypdf 
sentence-transformers
langchain[docarray]
docarray
openai
streamlit
tiktoken
openai
python-dotenv
langchain_community
langchain 